requirements.txt
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
cryptography==41.0.5
defusedxml==0.8.0rc2
Django==4.2.5
django-filter==23.3
django-phonenumber-field==7.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.2
fiona==1.9.5
geopandas==0.14.1
geographiclib==2.0
geopy==2.4.1
idna==3.4
importlib-metadata==7.0.0
networkx==3.2.1
numpy==1.26.2
oauthlib==3.2.2
osmnx==1.8.0
packaging==23.2
pandas==2.1.4
phonenumbers==8.13.25
Pillow==10.1.0
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
pyproj==3.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.3.post1
requests==2.31.0
requests-oauthlib==1.3.1
shapely==2.0.2
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.1.0
zipp==3.17.0
